+++ src/app/schemas/components/tables.scss
