+++ src/index.html
@@ -7,6 +7,7 @@
 
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <link rel="icon" type="image/x-icon" href="favicon.ico">
+  <base href="/">
 </head>
 <body>
   <app-root>Loading...</app-root>
